Changes interfaceDatabase.py to FileInterface.py, Added files to run from main, a view clas file, and moved model functions over to JsonModel.py
