--- src/interfaceDatabase.py
@@ -1,57 +1,0 @@
-import abc
-import json
-class InterfaceToFile(abc):
-    @abc.abstractclassmethod
-    def put(self):
-        pass
-
-    @abc.abstractclassmethod
-    def post(self):
-        pass
-
-    @abc.abstractclassmethod
-    def delete(self):
-        pass
-
-class InterfaceFromFile(abc):
-    @abc.abstractclassmethod
-    def get(self):
-        pass
-
-  
-class InterfaceJson(InterfaceToFile):
-    def get(self):
-        f = open('data.json')
-        data = json.load(f)
-        for i in data:
-            print(i)
-            f.close()
-
-class InterfaceJson(InterfaceFromFile):
-
-    def put(self, emailToInsert, campsiteToInsert, dateToInsert):
-        f = open('data.json')
-        data = json.load(f)
-        toAppend = {'email': emailToInsert,
-                'campsite': campsiteToInsert, 'date': dateToInsert}
-        data.append(toAppend)
-        outFile = open("data.json", "w")
-        json.dump(data, outFile)
-    
-    def delete(self, emailToFind):
-        f = open('data.json')
-        data = json.load(f)
-        for i, obj in enumerate(data):
-            if (obj["email"] == emailToFind):
-                listobj.pop(i)
-        outFile = open("updated.json", "w")
-        json.dump(listobj, outFile)
-
-    def findCampsite(self, campsiteToFind):
-        f = open('data.json')
-        data = json.load(f)
-        for i, obj in enumerate(data):
-            if (obj["campsite"] == campsiteToFind):
-                return True
-            else:
-                return False
